requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
anyio==4.4.0
argcomplete==1.10.3
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.8.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
compressed-rtf==1.0.6
configobj==5.0.8
configparser==7.0.0
Django==5.0.3
django-bootstrap5==23.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
docx2txt==0.8
easyocr==1.7.1
ebcdic==1.1.1
etelemetry==0.3.1
exceptiongroup==1.2.1
extract-msg==0.28.7
filelock==3.15.4
flatbuffers==24.3.25
frontend==0.0.3
fsspec==2024.6.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
gTTS==2.5.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
idna==3.6
imageio==2.34.2
IMAPClient==2.1.0
intel-openmp==2021.4.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
keras==3.3.2
lazy_loader==0.4
libclang==18.1.1
looseversion==1.3.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.3
nibabel==5.2.1
ninja==1.11.1.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
olefile==0.47
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pathlib==1.0.1
pdfminer.six==20191110
pillow==10.3.0
protobuf==4.25.3
prov==2.0.1
psycopg2-binary==2.9.9
PyAudio==0.2.14
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydot==2.0.0
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
PyPDF2==3.0.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-pptx==0.6.23
pytils==0.4.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==6.3.2
regex==2023.12.25
requests==2.31.0
rich==13.7.1
scikit-image==0.24.0
scikit-learn==1.4.2
scipy==1.13.0
shapely==2.0.4
simplejson==3.19.2
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
sqlparse==0.4.4
starlette==0.37.2
sympy==1.12.1
tbb==2021.13.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.6.18
tools==0.1.9
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.2
traits==6.3.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.30.1
Werkzeug==3.0.2
wikipedia==1.4.0
wrapt==1.16.0
xlrd==1.2.0
XlsxWriter==3.2.0
